chged cfg .assign.Rgr.remove.failures to G_INT16
plus:
 added Config::Property::Value::Type::G_UINT16
 added Config::Property::Value::V_GUINT16
 added ::SWC::Config::g_i16(const uint16_t& v)

[TEST COMMIT]

+++ src/cc/include/swcdb/core/comm/Resolver.h
@@ -17,8 +17,8 @@ namespace SWC {
 
 namespace Comm {
 
-typedef asio::ip::tcp::endpoint         EndPoint;
-typedef Core::Vector<EndPoint, uint8_t> EndPoints;
+typedef asio::ip::tcp::endpoint          EndPoint;
+typedef Core::Vector<EndPoint, uint32_t> EndPoints;
 
 struct Network {
   bool                        is_v4;
